moved gaia logic related types into their own Gaia.Types module
