+++ src/Gaia/Directives.hs
@@ -6,6 +6,8 @@ module Gaia.Directives(
   parseDirectivesFile
 ) where
 
+import Gaia.Types
+
 import           Control.Monad          (fail)
 
 import           Text.Parsec            (alphaNum, char, many, newline, parse,
@@ -19,18 +21,6 @@ import           Text.Parsec.String     (Parser, parseFromFile)
 -- NOTES
 -- - to print UTF-8 character we cannot use putStr but instead we
 --   will need to use Text or `mapM_ putChar`
-
-
--- SUPPORT TYPES
-type DirectiveBody = String
-data DirectiveTag  = Tag        -- | NewTag1 | NewTag2 ...
-                     deriving (Eq, Show)
-
-data Directive = Directive DirectiveTag DirectiveBody
-                 deriving (Eq)
-
-instance Show Directive where
-  show (Directive t b) = show t ++ " -> " ++ "\"" ++ b ++ "\""
 
 
 -- HELPERS
